refactor(gradle): Move `subprojects` code to conventions
This continues the effort to reduce implicit project coupling.
